+++ .dist-env
@@ -1,0 +1,3 @@
+PERSPECTIVE_API_KEY=
+DISCORD_API_KEY=
+DISCOVERY_URL=https://commentanalyzer.googleapis.com/$discovery/rest?version=v1alpha1
