+++ .gitignore
@@ -2,7 +2,7 @@
 ## files generated by popular Visual Studio add-ons.
 ##
 ## Get latest from https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
-
+.env
 # User-specific files
 *.rsuser
 *.suo
@@ -349,3 +349,5 @@ MigrationBackup/
 # Ionide (cross platform F# VS Code tools) working folder
 .ionide/
 src/.env
+
+/dist
