+++ package.json
@@ -1,11 +1,12 @@
 {
   "name": "batur.ai",
-  "type": "module",
+  "type": "commonjs",
   "version": "1.0.0",
   "description": "Bad Actor and Toxic User Removal is a Discord bot developed by IBXCODECAT! BATUR uses AI to scan for \"bad faith acting\" and high levels of toxicity within a discord message. If BATUR flags a message, it will perform a custom moderation action specified by the server administrators.",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc -p .",
+    "start": "node dist/index.js"
   },
   "repository": {
     "type": "git",
@@ -20,8 +21,10 @@
   "dependencies": {
     "discord.js": "^13.6.0",
     "dotenv": "^16.0.0",
-    "fs": "^0.0.1-security",
-    "googleapis": "^46.0.0",
-    "os": "^0.1.2"
+    "googleapis": "^95.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "^17.0.15",
+    "typescript": "^4.5.5"
   }
 }
