+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "display": "Node 16",
+  
+    "compilerOptions": {
+      "lib": ["es2021"],
+      "module": "commonjs",
+      "target": "es2021",
+      "outDir": "dist",
+  
+      "strict": true,
+      "skipLibCheck": true,
+      "forceConsistentCasingInFileNames": true
+    }
+  }
